+++ src/Container/ApplicationFactory.php
@@ -1,34 +1,33 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/mezzio/mezzio for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Expressive\Container;
+namespace Mezzio\Container;
 
 use Interop\Container\ContainerInterface;
-use Zend\Diactoros\Response\EmitterInterface;
-use Zend\Diactoros\Response\SapiEmitter;
-use Zend\Expressive\Application;
-use Zend\Expressive\Emitter\EmitterStack;
-use Zend\Expressive\Exception;
-use Zend\Expressive\Router\Aura as AuraRouter;
-use Zend\Expressive\Router\Route;
-use Zend\Expressive\Router\RouterInterface;
+use Laminas\Diactoros\Response\EmitterInterface;
+use Laminas\Diactoros\Response\SapiEmitter;
+use Mezzio\Application;
+use Mezzio\Emitter\EmitterStack;
+use Mezzio\Exception;
+use Mezzio\Router\Aura as AuraRouter;
+use Mezzio\Router\Route;
+use Mezzio\Router\RouterInterface;
 
 /**
  * Factory to use with an IoC container in order to return an Application instance.
  *
  * This factory uses the following services, if available:
  *
- * - 'Zend\Expressive\Router\RouterInterface'. If missing, an Aura router bridge
+ * - 'Mezzio\Router\RouterInterface'. If missing, an Aura router bridge
  *   will be instantiated and used.
- * - 'Zend\Expressive\FinalHandler'. The service should be a callable to use as
+ * - 'Mezzio\FinalHandler'. The service should be a callable to use as
  *   the final handler when the middleware pipeline is exhausted.
- * - 'Zend\Diactoros\Response\EmitterInterface'. If missing, an EmitterStack is
+ * - 'Laminas\Diactoros\Response\EmitterInterface'. If missing, an EmitterStack is
  *   created, adding a SapiEmitter to the bottom of the stack.
  * - 'config' (an array or ArrayAccess object). If present, and it contains route
  *   definitions, these will be used to seed routes in the Application instance
@@ -58,7 +57,7 @@ use Zend\Expressive\Router\RouterInterface;
  * Each route MUST have a path and middleware key at the minimum.
  *
  * The "allowed_methods" key may be omitted, can be either an array or the
- * value of the Zend\Expressive\Router\Route::HTTP_METHOD_ANY constant; any
+ * value of the Mezzio\Router\Route::HTTP_METHOD_ANY constant; any
  * valid HTTP method token is allowed, which means you can specify custom HTTP
  * methods as well.
  *
@@ -99,7 +98,7 @@ use Zend\Expressive\Router\RouterInterface;
  *
  * Note that the `path` element can only be a literal. `error` indicates
  * whether or not the middleware represents error middleware; this is done
- * so that Expressive can lazy-load an error middleware service (more below).
+ * so that Mezzio can lazy-load an error middleware service (more below).
  * Omitting `error` or setting it to a non-true value is the default,
  * indicating the middleware is standard middleware.
  *
@@ -126,15 +125,19 @@ class ApplicationFactory
     {
         $router = $container->has(RouterInterface::class)
             ? $container->get(RouterInterface::class)
-            : new AuraRouter();
+            : ($container->has(\Zend\Expressive\Router\RouterInterface::class)
+                ? $container->get(\Zend\Expressive\Router\RouterInterface::class)
+                : new AuraRouter());
 
-        $finalHandler = $container->has('Zend\Expressive\FinalHandler')
-            ? $container->get('Zend\Expressive\FinalHandler')
+        $finalHandler = $container->has('Mezzio\FinalHandler')
+            ? $container->get('Mezzio\FinalHandler')
             : null;
 
         $emitter = $container->has(EmitterInterface::class)
             ? $container->get(EmitterInterface::class)
-            : null;
+            : ($container->has(\Zend\Diactoros\Response\EmitterInterface::class)
+                ? $container->get(\Zend\Diactoros\Response\EmitterInterface::class)
+                : null);
 
         $app = new Application($router, $container, $finalHandler, $emitter);