+++ test/ApplicationTest.php
@@ -1,25 +1,24 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/mezzio/mezzio for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Expressive;
+namespace MezzioTest;
 
 use Interop\Container\ContainerInterface;
+use Laminas\Diactoros\Response\SapiEmitter;
+use Laminas\Diactoros\ServerRequest as Request;
+use Laminas\Stratigility\Route as StratigilityRoute;
+use Mezzio\Application;
+use Mezzio\Emitter\EmitterStack;
+use Mezzio\Router\Route;
+use Mezzio\Router\RouteResult;
 use PHPUnit_Framework_TestCase as TestCase;
 use Prophecy\Argument;
 use ReflectionProperty;
-use Zend\Diactoros\Response\SapiEmitter;
-use Zend\Diactoros\ServerRequest as Request;
-use Zend\Expressive\Application;
-use Zend\Expressive\Emitter\EmitterStack;
-use Zend\Expressive\Router\Route;
-use Zend\Expressive\Router\RouteResult;
-use Zend\Stratigility\Route as StratigilityRoute;
 
 class ApplicationTest extends TestCase
 {
@@ -28,7 +27,7 @@ class ApplicationTest extends TestCase
         $this->noopMiddleware = function ($req, $res, $next) {
         };
 
-        $this->router = $this->prophesize('Zend\Expressive\Router\RouterInterface');
+        $this->router = $this->prophesize('Mezzio\Router\RouterInterface');
     }
 
     public function getApp()
@@ -50,13 +49,13 @@ class ApplicationTest extends TestCase
     public function testConstructorAcceptsRouterAsAnArgument()
     {
         $app = $this->getApp();
-        $this->assertInstanceOf('Zend\Expressive\Application', $app);
+        $this->assertInstanceOf('Mezzio\Application', $app);
     }
 
     public function testApplicationIsAMiddlewarePipe()
     {
         $app = $this->getApp();
-        $this->assertInstanceOf('Zend\Stratigility\MiddlewarePipe', $app);
+        $this->assertInstanceOf('Laminas\Stratigility\MiddlewarePipe', $app);
     }
 
     public function testRouteMethodReturnsRouteInstance()
@@ -115,7 +114,7 @@ class ApplicationTest extends TestCase
     public function testCallingRouteWithOnlyAPathRaisesAnException()
     {
         $app = $this->getApp();
-        $this->setExpectedException('Zend\Expressive\Exception\InvalidArgumentException');
+        $this->setExpectedException('Mezzio\Exception\InvalidArgumentException');
         $app->route('/path');
     }
 
@@ -140,7 +139,7 @@ class ApplicationTest extends TestCase
     public function testCallingRouteWithAnInvalidPathTypeRaisesAnException($path)
     {
         $app = $this->getApp();
-        $this->setExpectedException('Zend\Expressive\Exception\InvalidArgumentException');
+        $this->setExpectedException('Mezzio\Exception\InvalidArgumentException');
         $app->route($path, 'middleware');
     }
 
@@ -207,7 +206,7 @@ class ApplicationTest extends TestCase
 
         $this->assertCount(1, $pipeline);
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $test  = $route->handler;
 
         $routeMiddleware = [$app, 'routeMiddleware'];
@@ -228,7 +227,7 @@ class ApplicationTest extends TestCase
 
         $this->assertCount(1, $pipeline);
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $test  = $route->handler;
 
         $this->assertSame($routeMiddleware, $test);
@@ -275,7 +274,7 @@ class ApplicationTest extends TestCase
 
     public function testAllowsInjectingEmitterAtInstantiation()
     {
-        $emitter = $this->prophesize('Zend\Diactoros\Response\EmitterInterface');
+        $emitter = $this->prophesize('Laminas\Diactoros\Response\EmitterInterface');
         $app     = new Application(
             $this->router->reveal(),
             null,
@@ -296,7 +295,7 @@ class ApplicationTest extends TestCase
             return $finalResponse;
         };
 
-        $emitter = $this->prophesize('Zend\Diactoros\Response\EmitterInterface');
+        $emitter = $this->prophesize('Laminas\Diactoros\Response\EmitterInterface');
         $emitter->emit(
             Argument::type('Psr\Http\Message\ResponseInterface')
         )->shouldBeCalled();
@@ -345,7 +344,7 @@ class ApplicationTest extends TestCase
         $this->assertCount(1, $pipeline);
 
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $this->assertSame([$app, 'routeMiddleware'], $route->handler);
         $this->assertEquals('/', $route->path);
     }
@@ -416,7 +415,7 @@ class ApplicationTest extends TestCase
         $pipeline = $r->getValue($app);
 
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $handler = $route->handler;
 
         $this->assertEquals('invoked', $handler('foo', 'bar'));
@@ -443,7 +442,7 @@ class ApplicationTest extends TestCase
         $pipeline = $r->getValue($app);
 
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $handler = $route->handler;
 
         $this->assertEquals('invoked', $handler('foo', 'bar', 'baz', 'bat'));
@@ -470,7 +469,7 @@ class ApplicationTest extends TestCase
         $pipeline = $r->getValue($app);
 
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $handler = $route->handler;
 
         $this->assertEquals('invoked', $handler('foo', 'bar'));
@@ -497,7 +496,7 @@ class ApplicationTest extends TestCase
         $pipeline = $r->getValue($app);
 
         $route = $pipeline->dequeue();
-        $this->assertInstanceOf('Zend\Stratigility\Route', $route);
+        $this->assertInstanceOf('Laminas\Stratigility\Route', $route);
         $handler = $route->handler;
 
         $this->assertEquals('invoked', $handler('foo', 'bar', 'baz', 'bat'));