+++ test/Container/TemplatedErrorHandlerFactoryTest.php
@@ -1,18 +1,17 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/mezzio/mezzio for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Expressive\Container;
+namespace MezzioTest\Container;
 
+use Mezzio\Container\TemplatedErrorHandlerFactory;
+use Mezzio\Template\TemplateInterface;
+use Mezzio\TemplatedErrorHandler;
 use PHPUnit_Framework_TestCase as TestCase;
-use Zend\Expressive\Container\TemplatedErrorHandlerFactory;
-use Zend\Expressive\Template\TemplateInterface;
-use Zend\Expressive\TemplatedErrorHandler;
 
 class TemplatedErrorHandlerFactoryTest extends TestCase
 {
@@ -25,6 +24,7 @@ class TemplatedErrorHandlerFactoryTest extends TestCase
     public function testReturnsATemplatedErrorHandler()
     {
         $this->container->has(TemplateInterface::class)->willReturn(false);
+        $this->container->has(\Zend\Expressive\Template\TemplateInterface::class)->willReturn(false);
         $this->container->has('config')->willReturn(false);
 
         $factory = $this->factory;
@@ -47,11 +47,12 @@ class TemplatedErrorHandlerFactoryTest extends TestCase
 
     public function testWillInjectTemplateNamesFromConfigurationWhenPresent()
     {
-        $config = ['zend-expressive' => ['error_handler' => [
+        $config = ['mezzio' => ['error_handler' => [
             'template_404'   => 'error::404',
             'template_error' => 'error::500',
         ]]];
         $this->container->has(TemplateInterface::class)->willReturn(false);
+        $this->container->has(\Zend\Expressive\Template\TemplateInterface::class)->willReturn(false);
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn($config);